Merge implementations and modernize
This commit does many things, some might say too many.  First it drops
the bazel build system and the Makefile, both of which are not very
standard within the Go ecosystem.  Bazel is also just a pain to deal
with.  As a continuation of this the vendor directory is dropped, as
this information is encoded in the go modules.

The go module setup is bootstrapped now using the data that was being
used by dep, future work is to update these dependencies to later
versions to take advantage of updates to linuxkit and its
dependencies.

The metadata data source is also now integrated with the provider.